rake double to float

+++ db/migrate/20160410064218_add_column_to_price_newest.rb
@@ -1,5 +1,5 @@
 class AddColumnToPriceNewest < ActiveRecord::Migration
   def change
-    add_column :price_newests, :previoustrade, :double
+    add_column :price_newests, :previoustrade, :float
   end
 end
